refactor: theme and component
