+++ packages/components/skeleton/package.json
@@ -37,6 +37,7 @@
   },
   "devDependencies": {
     "@chakra-ui/shared-utils": "workspace:*",
+    "@chakra-ui/object-utils": "workspace:*",
     "@chakra-ui/system": "workspace:*",
     "@chakra-ui/layout": "workspace:*",
     "@chakra-ui/theme": "workspace:*",
