+++ packages/components/skeleton/src/index.ts
@@ -1,1 +1,3 @@
 export * from "./skeleton"
+export * from "./skeleton-text"
+export * from "./skeleton-circle"
