+++ packages/components/skeleton/src/skeleton.tsx
@@ -1,16 +1,18 @@
-import { useBreakpointValue } from "@chakra-ui/media-query"
+import { compact } from "@chakra-ui/object-utils"
+import { usePrevious } from "@chakra-ui/react-use-previous"
+import { cx } from "@chakra-ui/shared-utils"
 import {
   chakra,
+  cssVar,
   forwardRef,
+  HTMLChakraProps,
   keyframes,
   omitThemingProps,
   ThemingProps,
   useStyleConfig,
-  HTMLChakraProps,
+  useToken,
 } from "@chakra-ui/system"
-import { usePrevious } from "@chakra-ui/react-use-previous"
-import { cx } from "@chakra-ui/shared-utils"
-import { useEffect, useRef } from "react"
+import { useIsFirstRender } from "./use-is-first-render"
 
 export interface SkeletonOptions {
   /**
@@ -54,15 +56,8 @@ const StyledSkeleton = chakra("div", {
   },
 })
 
-const useIsFirstRender = () => {
-  const isFirstRender = useRef(true)
-
-  useEffect(() => {
-    isFirstRender.current = false
-  }, [])
-
-  return isFirstRender.current
-}
+const $startColor = cssVar("skeleton-start-color")
+const $endColor = cssVar("skeleton-end-color")
 
 export type ISkeleton = SkeletonOptions
 
@@ -76,12 +71,15 @@ const fade = keyframes({
   to: { opacity: 1 },
 })
 
-const startColor = "var(--skeleton-start-color)"
-const endColor = "var(--skeleton-end-color)"
-
 const bgFade = keyframes({
-  from: { borderColor: startColor, background: startColor },
-  to: { borderColor: endColor, background: endColor },
+  from: {
+    borderColor: $startColor.reference,
+    background: $startColor.reference,
+  },
+  to: {
+    borderColor: $endColor.reference,
+    background: $endColor.reference,
+  },
 })
 
 /**
@@ -94,8 +92,8 @@ export const Skeleton = forwardRef<SkeletonProps, "div">((props, ref) => {
   const isFirstRender = useIsFirstRender()
 
   const {
-    startColor,
-    endColor,
+    startColor = "",
+    endColor = "",
     isLoaded,
     fadeDuration,
     speed,
@@ -103,9 +101,19 @@ export const Skeleton = forwardRef<SkeletonProps, "div">((props, ref) => {
     ...rest
   } = omitThemingProps(props)
 
+  const [startColorVar, endColorVar] = useToken("colors", [
+    startColor,
+    endColor,
+  ])
+
   const wasPreviouslyLoaded = usePrevious(isLoaded)
 
   const _className = cx("chakra-skeleton", className)
+
+  const cssVarStyles = compact({
+    [$startColor.variable]: startColorVar || undefined,
+    [$endColor.variable]: endColorVar || undefined,
+  })
 
   if (isLoaded) {
     const animation =
@@ -128,6 +136,8 @@ export const Skeleton = forwardRef<SkeletonProps, "div">((props, ref) => {
       {...rest}
       __css={{
         ...styles,
+        ...cssVarStyles,
+        _dark: { ...cssVarStyles },
         animation: `${speed}s linear infinite alternate ${bgFade}`,
       }}
     />
@@ -140,105 +150,3 @@ Skeleton.defaultProps = {
 }
 
 Skeleton.displayName = "Skeleton"
-
-function range(count: number) {
-  return Array(count)
-    .fill(1)
-    .map((_, index) => index + 1)
-}
-
-export interface SkeletonTextProps extends SkeletonProps {
-  spacing?: SkeletonProps["margin"]
-  skeletonHeight?: SkeletonProps["height"]
-  startColor?: SkeletonProps["startColor"]
-  endColor?: SkeletonProps["endColor"]
-  isLoaded?: SkeletonProps["isLoaded"]
-}
-
-const defaultNoOfLines = 3
-
-/**
- * `SkeletonText` is used to display the loading state in the form of text.
- *
- * @see Docs https://chakra-ui.com/docs/components/skeleton
- */
-export const SkeletonText: React.FC<SkeletonTextProps> = (props) => {
-  const {
-    noOfLines = defaultNoOfLines,
-    spacing = "0.5rem",
-    skeletonHeight = "0.5rem",
-    className,
-    startColor,
-    endColor,
-    isLoaded,
-    fadeDuration,
-    speed,
-    children,
-    ...rest
-  } = props
-
-  const noOfLinesValue =
-    useBreakpointValue(
-      typeof noOfLines === "number" ? [noOfLines] : noOfLines,
-    ) || defaultNoOfLines
-  const numbers = range(noOfLinesValue)
-
-  const getWidth = (index: number) => {
-    if (noOfLinesValue > 1) {
-      return index === numbers.length ? "80%" : "100%"
-    }
-    return "100%"
-  }
-
-  const _className = cx("chakra-skeleton__group", className)
-
-  return (
-    <chakra.div className={_className} {...rest}>
-      {numbers.map((number, index) => {
-        if (isLoaded && index > 0) {
-          // skip other lines
-          return null
-        }
-
-        const sizeProps = isLoaded
-          ? null
-          : {
-              mb: number === numbers.length ? "0" : spacing,
-              width: getWidth(number),
-              height: skeletonHeight,
-            }
-
-        return (
-          <Skeleton
-            key={numbers.length.toString() + number}
-            startColor={startColor}
-            endColor={endColor}
-            isLoaded={isLoaded}
-            fadeDuration={fadeDuration}
-            speed={speed}
-            {...sizeProps}
-          >
-            {
-              // allows animating the children
-              index === 0 ? children : undefined
-            }
-          </Skeleton>
-        )
-      })}
-    </chakra.div>
-  )
-}
-
-SkeletonText.displayName = "SkeletonText"
-
-/**
- * `SkeletonCircle` is used to display the loading state in the form of a circular avatar.
- *
- * @see Docs https://chakra-ui.com/docs/components/skeleton
- */
-export const SkeletonCircle: React.FC<SkeletonProps> = ({
-  size = "2rem",
-  ...rest
-}) => <Skeleton borderRadius="full" boxSize={size} {...rest} />
-
-SkeletonCircle.displayName = "SkeletonCircle"
