+++ packages/components/skeleton/src/use-is-first-render.ts
@@ -1,0 +1,11 @@
+import { useEffect, useRef } from "react"
+
+export function useIsFirstRender() {
+  const isFirstRender = useRef(true)
+
+  useEffect(() => {
+    isFirstRender.current = false
+  }, [])
+
+  return isFirstRender.current
+}
